chore(crowdsale): factor crowdsale out into SFC
